requirements.txt
Django==2.2
django-choices==1.6.2
django-filter==2.1.0
djangorestframework==3.9.2
factory-boy==2.11.1
jupyter==1.0.0
